+++ editor/js/helpers.js
@@ -4,3 +4,10 @@ function escHtml(s) {
 function escAttr(s) {
   return String(s).replace(/&/g,'&amp;').replace(/"/g,'&quot;').replace(/'/g,'&#39;');
 }
+
+function emptyPreviewHtml() {
+  return '<div class="preview-placeholder" id="preview-placeholder">'
+    + '<span>Start typing to render your diagram</span>'
+    + '<button class="placeholder-example-btn" type="button" data-action="load-example">Load an example</button>'
+    + '</div>';
+}